fixed 64-bits warnings
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1583 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/seas/encode_header.c
@@ -210,7 +210,7 @@ int encode_header(struct sip_msg *sipmsg,struct hdr_field *hdr,unsigned char *pa
 	       goto error;
 	    }
 	 }
-	 if((len=encode_content_type(hdr->name.s,hdr->len,(unsigned int)hdr->parsed,(char*)&payload[4]))<0){
+	 if((len=encode_content_type(hdr->name.s,hdr->len,(unsigned int)(unsigned long)hdr->parsed,(char*)&payload[4]))<0){
 	    myerror="ERROR: encode_header: encoding via header\n";
 	    goto error;
 	 }else{
